requirements-docker.txt
# Minimal-ish runtime deps for shape-only Hunyuan3D-2.1 worker.
# Pinned to reduce "latest broke" surprises.

fastapi==0.115.12
uvicorn==0.34.3
python-multipart==0.0.22
requests==2.32.5

numpy==1.26.4
pillow==10.4.0
opencv-python==4.10.0.84

huggingface-hub==0.30.2
safetensors==0.4.4
transformers==4.46.0
diffusers==0.30.0
accelerate==1.1.1
einops==0.8.0

rembg==2.0.65
onnxruntime==1.16.3

trimesh==4.4.7
pygltflib==1.16.3

# Needed by hy3dshape hunyuandit MoE layers
timm==0.9.16

# hy3dshape postprocessors import pymeshlab unconditionally
pymeshlab==2022.2.post3

